be/requirements.txt
pymongo==3.3.0
pyserial==3.2.1
redis==2.10.5
tornado==4.4.2